backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
firebase-admin>=6.4.0
httpx>=0.27.0
cryptography>=42.0.0
